agent/requirements.txt
# AI Agent Framework
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0

# Blockchain
web3>=6.15.0
eth-account>=0.11.0
eth-abi>=5.0.0

# Data & Analysis
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# IPFS
ipfshttpclient>=0.8.0a2

# Utilities
rich>=13.7.0
structlog>=24.1.0
